+++ scripts/check-versions.mjs
@@ -14,6 +14,7 @@ const packages = [
   'core',
   'dom',
   'jsx-runtime',
+  'deep',
   'router',
   'query',
   'styled',